Removing dependency on Net::DNS::ZoneFile::Fast.
Net::DNS::ZoneFile::Fast is a third-party module -- not part of Net::DNS.
When yazvs was first written, it was the only module that could parse
a zone file.  Now Net::DNS has a built-in Net::DNS::ZoneFile package
that seems to work well, so we can remove this one dependency.